Add hooks when autoloading constants
Add the ability to provide a block when a constant is autoloaded. This
is very useful when using the popular pattern to seperate class
definition and configuration. Frequently we'll see `MyClass.configure`
in an initializer, this causes eager loading of the MyClass class,
slowing down boot time, instead what we can now do with this patch is to
lazily load the configuration when the constant is autoloaded using:

```
ActiveSupport::Dependencies.onload('MyClass') do
  MyClass.configure do |config|
    # config here ...
  end
end
```

These blocks will be called each time the constant is loaded, for
example when a constant is autoreloaded due to modifying the file. The
nice side effect of this is we no longer have to use `config.to_prepare`
to ensure that our configuration is not lost.